+++ backend/goodday/src/main/java/seeuthere/goodday/auth/utils/Naver.java
@@ -1,4 +1,4 @@
-package seeuthere.goodday.auth.domain;
+package seeuthere.goodday.auth.utils;
 
 
 import java.math.BigInteger;
